refactor(interview): dedupe extractCustomAttributes and unexport internal parentage types
