+++ packages/interview/src/interfaces/FamilyPedigree/components/wizards/transforms/childCellTransform.ts
@@ -1,26 +1,11 @@
-import type { NcEdge, NcNode, VariableValue } from '@codaco/shared-consts';
+import type { NcEdge, NcNode } from '@codaco/shared-consts';
 import type {
   CommitBatch,
   VariableConfig,
 } from '~/interfaces/FamilyPedigree/store';
 
 import { buildChildParentage } from './buildChildParentage';
-
-const KNOWN_PERSON_KEYS = new Set(['name']);
-
-function extractCustomAttributes(
-  obj: Record<string, unknown>,
-): Record<string, VariableValue> | undefined {
-  const attrs: Record<string, VariableValue> = {};
-  let hasAttrs = false;
-  for (const [key, val] of Object.entries(obj)) {
-    if (!KNOWN_PERSON_KEYS.has(key) && val !== undefined) {
-      attrs[key] = val as VariableValue;
-      hasAttrs = true;
-    }
-  }
-  return hasAttrs ? attrs : undefined;
-}
+import { extractCustomAttributes } from './personAttributes';
 
 export function childCellTransform(
   values: Record<string, unknown>,
