+++ packages/interview/src/interfaces/FamilyPedigree/components/wizards/transforms/personAttributes.ts
@@ -1,0 +1,17 @@
+import type { VariableValue } from '@codaco/shared-consts';
+
+const KNOWN_PERSON_KEYS = new Set(['name']);
+
+export function extractCustomAttributes(
+  obj: Record<string, unknown>,
+): Record<string, VariableValue> | undefined {
+  const attrs: Record<string, VariableValue> = {};
+  let hasAttrs = false;
+  for (const [key, val] of Object.entries(obj)) {
+    if (!KNOWN_PERSON_KEYS.has(key) && val !== undefined) {
+      attrs[key] = val as VariableValue;
+      hasAttrs = true;
+    }
+  }
+  return hasAttrs ? attrs : undefined;
+}
