+++ packages/interview/src/interfaces/FamilyPedigree/components/wizards/transforms/siblingCellTransform.ts
@@ -4,21 +4,7 @@ import type {
   VariableConfig,
 } from '~/interfaces/FamilyPedigree/store';
 
-const KNOWN_PERSON_KEYS = new Set(['name']);
-
-function extractCustomAttributes(
-  obj: Record<string, unknown>,
-): Record<string, VariableValue> | undefined {
-  const attrs: Record<string, VariableValue> = {};
-  let hasAttrs = false;
-  for (const [key, val] of Object.entries(obj)) {
-    if (!KNOWN_PERSON_KEYS.has(key) && val !== undefined) {
-      attrs[key] = val as VariableValue;
-      hasAttrs = true;
-    }
-  }
-  return hasAttrs ? attrs : undefined;
-}
+import { extractCustomAttributes } from './personAttributes';
 
 function buildPersonAttributes(
   person: Record<string, unknown>,
